refactor(hub): co-locate hub data and game files, add blocked paths system
- Move hub JSON config to web/src/data/hub/ (config.json, questDefs.json)
- Consolidate hubConfigLoader.ts + 3 shim files into hub/loader.ts
- Move hub game logic to web/src/game/hub/ (quests, pickups, friendship, innConvos)
- Add blocked paths system: JSON-driven quest-gated obstructions with NPC guards,
  proximity speech bubbles, and decor in blocked/cleared states
- Add requireTouch flag on pickup items for tile-arrival collection
- Update all import paths in HubTownCanvas, HubWorld, QuestsModal, stories
- Add docs/hubworld.md — schema reference for blocked paths, pickup items, tile keys
- Add docs/hubworld.md reference to AGENTS.md
