+++ web/src/components/hub/HubWorld.tsx
@@ -5,7 +5,13 @@ import { AreaNameBadge } from './AreaNameBadge'
 import { HubReturnButton } from './HubReturnButton'
 import { HubDialogue } from './HubDialogue'
 import type { DialogueChoice } from './HubDialogue'
-import { AVATAR_START, MAP_W, MAP_H } from '../../data/hubLayout'
+import { AVATAR_START, MAP_W, MAP_H, HUB_NPCS } from '../../data/hub/loader'
+import { HUB_QUEST_DEFS, INN_RUMOURS, FRIENDSHIP_DIALOGUE } from '../../data/hub/questDefs'
+import type { HubQuestDef } from '../../data/hub/questDefs'
+import { getPickedUpIds, markPickedUp, unmarkPickedUp } from '../../game/hub/pickups'
+import { getFriendshipLevel, addFriendshipXp, getFriendshipData } from '../../game/hub/friendship'
+import { getQuestState, setQuestStatus, incrementQuestProgress, getQuestProgress, resetQuest } from '../../game/hub/quests'
+import { getHeardConvoIds, markConvoHeard } from '../../game/hub/innConvos'
 import { loadDeck, loadCollection, loadCrystals, saveCrystals } from '../../game/collection'
 import { getCardCatalog } from '../../game/cards'
 import { CommanderState } from '../../game/commander'
@@ -18,13 +24,6 @@ import { ToolbarSpacer } from '../ui/Toolbar/ToolbarSpacer'
 import { User } from 'firebase/auth'
 import { loadPlayerName } from '../../game/questline'
 import { LoginButton } from '../ui/LoginButton'
-import { HUB_NPCS } from '../../data/hubConfigLoader'
-import { HUB_QUEST_DEFS, INN_RUMOURS, FRIENDSHIP_DIALOGUE } from '../../data/hubQuestDefs'
-import type { HubQuestDef } from '../../data/hubQuestDefs'
-import { getPickedUpIds, markPickedUp, unmarkPickedUp } from '../../game/hubPickups'
-import { getFriendshipLevel, addFriendshipXp, getFriendshipData } from '../../game/hubFriendship'
-import { getQuestState, setQuestStatus, incrementQuestProgress, getQuestProgress, resetQuest } from '../../game/hubQuests'
-import { getHeardConvoIds, markConvoHeard } from '../../game/hubInnConvos'
 import { addCollectible, getCollectibles } from '../../game/itemStore'
 import { QuestsModal } from './QuestsModal'
