--- web/src/data/hubInteriors.ts
@@ -1,3 +1,0 @@
-// Hub World building interiors — thin re-export shim. Edit web/src/data/hubConfig.json to change interiors.
-export type { HubDoor, InteriorDecor, HubInterior } from './hubConfigLoader'
-export { HUB_DOORS, HUB_INTERIORS } from './hubConfigLoader'
