--- web/src/data/hubLayout.ts
@@ -1,3 +1,0 @@
-// Hub World layout — thin re-export shim. Edit web/src/data/hubConfig.json to change map data.
-export type { HubArea, HubBuilding, HubStreetGroup } from './hubConfigLoader'
-export { MAP_W, MAP_H, AVATAR_START, HUB_AREAS, HUB_STREET_TILES, HUB_STREET_GROUPS, HUB_BUILDING_TILES, HUB_BUILDINGS } from './hubConfigLoader'
