--- web/src/data/hubNpcs.ts
@@ -1,3 +1,0 @@
-// Hub NPC definitions — thin re-export shim. Edit web/src/data/hubConfig.json to change NPCs.
-export type { HubNpc } from './hubConfigLoader'
-export { HUB_NPCS, EXTERIOR_NPCS, INTERIOR_NPCS, NPC_SPAWN_TILES, AMBIENT_NPC_SPRITES } from './hubConfigLoader'
