+++ web/src/game/hub/friendship.ts
