+++ web/src/game/hub/pickups.ts
